Rename SUPPORTED_VIEWS to SUPPORTED_SYNONYMS
