+++ lib/supported-synonyms.js
